requirements.txt
annotated-types==0.7.0
anyio==4.12.1
-e git+ssh://git@github.com/alessandro54/transcript.git@27fb4f5e6d722a1e376a5a08d4e42a00f5453b0d#egg=audio_transcription_bot
av==16.1.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
ctranslate2==4.6.3
cycler==0.12.1
distro==1.9.0
dotenv==0.9.9
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
humanfriendly==10.0
idna==3.11
jiter==0.12.0
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
mpmath==1.3.0
mutagen==1.47.0
noisereduce==3.0.3
numpy==2.4.1
onnxruntime==1.23.2
openai==1.59.6
packaging==25.0
pillow==12.1.0
protobuf==6.33.4
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
PyYAML==6.0.3
scipy==1.17.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
